# task2/github_info_system/collector/internal/domain/repository.go
package domain

import "time"

type Repository struct {
	Name        string   
	Description string    
	Stars       int       
	Forks       int       
	CreatedAt   time.Time
}

type RepositoryInfo struct {
	Owner string
	Repo  string
}